chore: retire the dnaSharingMode knob preset and KnobTuningStrategy (#3554)
No consumer set `dnaSharingMode`, and its "default" preset was defined to
equal the per-knob defaults `createNeatConfig()` already applied, so the
whole preset layer was inert. The only non-default value, "aggressive",
measured zero lift in the #2496 bake-off.

- src/config/DnaSharingPreset.ts, src/transfer/KnobTuningStrategy.ts — gone.
- src/config/NeatConfig.ts — the four inter-island knob defaults are now
  inline literals; `mergeCompatibilityGatingDefaults` is gone because
  `parseCompatibilityGating` already applies the identical defaults.
- src/config/NeatArguments.ts, src/transfer/mod.ts — option and re-exports gone.
- bench/DnaSharingBakeOff.ts — the KnobTuning arm dropped; four arms remain.
- test/config/InterIslandKnobDefaults.ts — new, carries over the assertions
  that used to guard the preset/default equality.
- scripts/audit-option-usage.ts — negative control replaced with
  `syntheticAlignmentThreshold` (a KEEP key no consumer sets), verified
  against GRQ and NEAT-AI-Examples.
- Audit roll-up entry removed and pinned key count updated (113 -> 112).

Vibe-Coder-Run-Id: vibe-ms8i0nuj-e02144
