+++ docs/OPTION_AUDIT_CONSOLIDATED.md
@@ -42,7 +42,7 @@ grounds for removal.
 
 ```mermaid
 flowchart LR
-    H["#3518 harness<br/>enumerateOptionKeys()"] --> INV["267 rows<br/>113 top-level + 154 nested"]
+    H["#3518 harness<br/>enumerateOptionKeys()"] --> INV["266 rows<br/>112 top-level + 154 nested"]
     A["Slices A–F<br/>#3519–#3524"] --> TAB["Merged table<br/>optionAuditRollup.ts"]
     INV --> REC{"reconcile()"}
     TAB --> REC
@@ -62,7 +62,7 @@ run on every CI build by
 
 ```bash
 deno run --allow-read scripts/option-audit-rollup.ts
-# 🔎 267 enumerated rows (113 top-level, 154 nested) · 267 classified
+# 🔎 266 enumerated rows (112 top-level, 154 nested) · 266 classified
 # ✅ zero coverage gaps — every option key is classified
 ```
 
@@ -77,7 +77,7 @@ in #3518 and #3525. The original figure came from a grep that skipped
 `readonly mutation: readonly MutationInterface[]`; the parser-backed harness saw
 the real 119, and its CI test pins that number. Every slice brief was written
 from the 118-key list, so **no slice classified `mutation`**. (The pinned count
-is **113** today, as executed removals have taken keys back out. See
+is **112** today, as executed removals have taken keys back out. See
 [Executed removals](#executed-removals).)
 
 This roll-up classified it with the slice method — `git grep`, exit code
@@ -106,6 +106,7 @@ shrink as the campaign proceeds.
 | #3552 | `maxConns`, `maximumNumberOfNodes` | A     | Both `Mutator` growth guards gone; `maxSynapses` still caps.          |
 | #3553 | `enableRepetitiveTraining`         | A     | Repeat-training gate gone; the schedule guard is now unconditional.   |
 | #3559 | `novelty`                          | C     | Keep-or-remove decision resolved as remove; engine and 6 fields gone. |
+| #3554 | `dnaSharingMode`                   | A     | Knob preset and `KnobTuningStrategy` gone; knob defaults inlined.     |
 
 ### `fitnessSampleRate` and `seed`
 
@@ -127,19 +128,19 @@ per-slice totals shift:
 
 | Slice     | Slice comment | Roll-up | Why                                                                                                                                               |
 | --------- | ------------: | ------: | ------------------------------------------------------------------------------------------------------------------------------------------------- |
-| A         |            46 |      43 | —                                                                                                                                                 |
+| A         |            46 |      42 | —                                                                                                                                                 |
 | B         |            36 |      42 | B classified its 3 nested configs at parent level; the harness enumerates `DiscoveryCacheConfig` (4) + `DiskSpaceConfig` (3) fields individually. |
 | C         |            59 |      38 | `adaptiveMutationThresholds` is declared inline in `NeatArguments`, so its 3 fields are not separate nested rows.                                 |
 | D         |            69 |      63 | Same for `plateauDetection`'s 6 inline fields.                                                                                                    |
 | E         |            33 |      37 | `RustScorerConfig` has no `NeatOptions` key (−1 parent row) but two interfaces, so `RequiredRustScorerConfig`'s 5 fields are enumerated too (+5). |
 | F         |            43 |      43 | —                                                                                                                                                 |
 | roll-up   |             — |       1 | `mutation`.                                                                                                                                       |
-| **Total** |       **286** | **267** |                                                                                                                                                   |
+| **Total** |       **286** | **266** |                                                                                                                                                   |
 
 No slice lost a key in the merge; the deltas are enumeration conventions plus
-the [executed removals](#executed-removals) — A is down 3 (#3552 and #3553), B
-down 1 (#3556) and C down 18 (#3558 and its 10 nested fields, plus #3559 and its
-6 nested fields).
+the [executed removals](#executed-removals) — A is down 4 (#3552, #3553 and
+#3554), B down 1 (#3556) and C down 18 (#3558 and its 10 nested fields, plus
+#3559 and its 6 nested fields).
 
 ## Merged classification table
 
@@ -270,7 +271,6 @@ set independently. Generated by
 | `compatibilityGating`                         | C       | `KEEP`      | —     |                                                                                           |
 | `selectionPressure`                           | C       | `KEEP`      | —     |                                                                                           |
 | `tolerateCorruptParents`                      | A       | `KEEP`      | —     |                                                                                           |
-| `dnaSharingMode`                              | A       | `QUALIFIES` | #3554 | Slice F confirmed `DnaSharingPreset.ts` is covered here and filed nothing.                |
 | `memory.enabled`                              | E       | `IN USE`    | —     |                                                                                           |
 | `memory.proactiveGc`                          | E       | `QUALIFIES` | #3565 |                                                                                           |
 | `memory.nativeBudgetBytes`                    | E       | `IN USE`    | —     |                                                                                           |
@@ -291,7 +291,7 @@ each. None did — every one resolves to a single issue, verified by
 
 | Overlap                                 | Slices | Resolution                                                                                                            |
 | --------------------------------------- | ------ | --------------------------------------------------------------------------------------------------------------------- |
-| `dnaSharingMode` / `DnaSharingPreset`   | A ↔ F  | Slice A filed **#3554**; slice F recorded the overlap and filed nothing. One issue.                                   |
+| `dnaSharingMode` / `DnaSharingPreset`   | A ↔ F  | Slice A filed **#3554**; slice F recorded the overlap and filed nothing. One issue, now landed.                       |
 | `seed` / `rng`                          | A ↔ E  | Neither qualifies, so no issue exists to duplicate. `rng` is `KEEP` (slice E); `seed` is not a `NeatArguments` field. |
 | `discoveryCache` / `discoveryDiskSpace` | B ↔ E  | Both are `KEEP` on slice B's evidence; slice E did not reclassify them. No issue exists.                              |
