+++ src/config/NeatConfig.ts
@@ -80,12 +80,6 @@ import {
   resolveDiscoveryWorkerThreadCap,
 } from "@config/DiscoveryWorkerEnvelope.ts";
 
-// Issue #2492: DNA-sharing knob preset
-import {
-  type DnaSharingMode,
-  getDnaSharingPreset,
-} from "@config/DnaSharingPreset.ts";
-
 /**
  * Default cost of growth value used when not specified in options.
  */
@@ -163,35 +157,6 @@ export const DEFAULT_HEAVY_TASK_WORKER_COUNT = 2;
  * Provides a read-only configuration object for the NEAT algorithm.
  */
 export type NeatConfig = Readonly<NeatArguments>;
-
-/**
- * Merge per-knob defaults from a DNA-sharing preset into the user's
- * `compatibilityGating` overrides (Issue #2492). Returns a new object
- * where preset values fill in any field the user did not supply, so the
- * downstream `parseCompatibilityGating` parser sees mode-aware defaults.
- *
- * Internal helper — exported only for unit testing the merge semantics.
- */
-function mergeCompatibilityGatingDefaults(
-  userOverrides: Record<string, unknown> | undefined,
-  presetDefaults: { enabled: boolean; power: number; maxDraws: number },
-): Record<string, unknown> {
-  const merged: Record<string, unknown> = {
-    enabled: presetDefaults.enabled,
-    power: presetDefaults.power,
-    maxDraws: presetDefaults.maxDraws,
-  };
-  if (userOverrides) {
-    if (userOverrides.enabled !== undefined) {
-      merged.enabled = userOverrides.enabled;
-    }
-    if (userOverrides.power !== undefined) merged.power = userOverrides.power;
-    if (userOverrides.maxDraws !== undefined) {
-      merged.maxDraws = userOverrides.maxDraws;
-    }
-  }
-  return merged;
-}
 
 /**
  * Creates a validated NEAT configuration from user options.
@@ -288,15 +253,6 @@ export function createNeatConfig(options: NeatOptionsInput): NeatConfig {
     populationSize,
     costName,
   );
-
-  // Issue #2492: Resolve the DNA-sharing knob preset before parsing the
-  // five inter-island knobs so the preset can supply mode-aware defaults.
-  // Explicit user values still win over the preset default — `parseNumber`
-  // only reaches the default when `opts.<knob>` is undefined.
-  const dnaSharingMode: DnaSharingMode = opts.dnaSharingMode === "aggressive"
-    ? "aggressive"
-    : "default";
-  const dnaPreset = getDnaSharingPreset(dnaSharingMode);
 
   const config: NeatArguments = {
     creativeThinkingConnectionCount: parseNumber(
@@ -468,7 +424,7 @@ export function createNeatConfig(options: NeatOptionsInput): NeatConfig {
     diversityBreedingRate: parseNumber(
       "Diversity breeding rate",
       opts.diversityBreedingRate,
-      dnaPreset.diversityBreedingRate,
+      0,
       { min: 0, max: 1 },
     ),
     CRISPRs: options.CRISPRs || [],
@@ -487,13 +443,13 @@ export function createNeatConfig(options: NeatOptionsInput): NeatConfig {
     geneticCompatibilityThreshold: parseNumber(
       "Genetic Compatibility Threshold",
       opts.geneticCompatibilityThreshold,
-      dnaPreset.geneticCompatibilityThreshold,
+      0.3,
       { min: 0, max: 1 },
     ),
     interSpeciesCrossoverThreshold: parseNumber(
       "Inter-species Crossover Threshold",
       opts.interSpeciesCrossoverThreshold,
-      dnaPreset.interSpeciesCrossoverThreshold,
+      0.1,
       { min: 0, max: 1 },
     ),
     syntheticAlignmentThreshold: parseNumber(
@@ -766,14 +722,8 @@ export function createNeatConfig(options: NeatOptionsInput): NeatConfig {
       opts.speciesStagnation as Record<string, unknown> | undefined,
     ),
     // Issue #2455: Parse compatibility gating configuration
-    // Issue #2492: Apply DNA-sharing preset defaults to any field the user
-    // did not set explicitly. The preset overrides only the *defaults*;
-    // user-supplied values still take precedence.
     compatibilityGating: parseCompatibilityGating(
-      mergeCompatibilityGatingDefaults(
-        opts.compatibilityGating as Record<string, unknown> | undefined,
-        dnaPreset.compatibilityGating,
-      ),
+      opts.compatibilityGating as Record<string, unknown> | undefined,
     ),
 
     // Issue #2929: Parse selection-pressure configuration (POWER exponent,
@@ -781,9 +731,6 @@ export function createNeatConfig(options: NeatOptionsInput): NeatConfig {
     selectionPressure: parseSelectionPressure(
       opts.selectionPressure as Record<string, unknown> | undefined,
     ),
-
-    // Issue #2492: Knob-tuning preset for inter-island DNA sharing.
-    dnaSharingMode,
 
     // Issue #2523: Tolerate corrupt parents during breeding by default.
     // Setting `false` restores legacy fail-fast behaviour.
